refactor: change the data flow from string to slice of integers

+++ domain/formatter/groups_of_thousands.go
@@ -15,7 +15,7 @@ const max = 9999
 const min = -9999
 const thousandCommasFrequency = 3
 
-func (f Formatter) GroupsOfThousands(number int, format bool) string {
+func (f Formatter) GroupsOfThousands(number int64, format bool) string {
 	strNumber := fmt.Sprint(number)
 	if format && (number > max || number < min) {
 		for i := len(strNumber) - thousandCommasFrequency; i > 0; i -= thousandCommasFrequency {
